+++ README.MD
@@ -43,5 +43,4 @@ Minor todo's:
  - probably kmz layout can be improved for better appearance in GE's file explorer
 
 Known issues:
- - linking tiles doesn't link like it should. It only checks underlying level. (Level 9 attempts to link Level 11, but not Level 13)
- If there would be missing tile in between, linking will skip all underlying tiles, thus resulting kmz will not contain all data.
+ - (SOLEVD) linking tiles doesn't link like it should. It only checks underlying level. (Level 9 attempts to link Level 11, but not Level 13) If there would be missing tile in between, linking will skip all underlying tiles, thus resulting kmz will not contain all data.
