+++ src/common/output.py
