+++ src/kml.py
@@ -1,5 +1,7 @@
+import src.dump as dump
+
+import xml.etree.ElementTree
 import os.path
-import xml.etree.ElementTree;
 
 def AppendData(lhs:xml.etree.ElementTree, rhs:xml.etree.ElementTree):
     ns = {'ns':'http://www.opengis.net/kml/2.2'}
@@ -8,13 +10,6 @@ def AppendData(lhs:xml.etree.ElementTree, rhs:xml.etree.ElementTree):
     if folder.find('name').text == 'Lines':
         for line in rhs.findall(".//Placemark", ns):
             folder.append(line)
-
-        # for child in folder:
-        #     print(child.tag, child.attrib, child.text)
-        # print("Append data...")
-        # for line in rhs.findall(".//Placemark", ns):
-        #     folder.append(line)
-    # print('\n')
 
     return lhs
 
@@ -26,3 +21,17 @@ def WriteDownKmlTree(root:xml.etree.ElementTree, output_file):
     else:
         tree = xml.etree.ElementTree.ElementTree(root)
     tree.write(output_file)
+
+def CreateTree(tile_id, lines, bbox, tile_level):
+
+    root = xml.etree.ElementTree.Element("kml")
+    doc = xml.etree.ElementTree.SubElement(root, "Document")
+    xml.etree.ElementTree.SubElement(doc, "name").text = "L" + str(tile_level) + ':' + str(tile_id.x) + ':' + str(tile_id.y)
+
+    styles = xml.etree.ElementTree.parse("resources/styles.xml")
+    for element in styles.getroot().findall(".//"):
+        doc.append(element)
+
+    dump.Tile(doc, bbox, tile_id, lines, tile_level)
+
+    return root
